Completed react-context optional refactor.
