+++ react-context/my-app/src/App.tsx
@@ -4,20 +4,11 @@ import { About } from './pages/About';
 import { Catalog } from './pages/Catalog';
 import { NotFound } from './pages/NotFound';
 import { ProductDetails } from './pages/ProductDetails';
-import { CartContext } from './components/CartContext';
-import { useState } from 'react';
-import { Product } from './lib';
+import { CartProvider } from './components/CartContext';
 
 export function App() {
-  const [cart, setCart] = useState<Product[]>([]);
-  const addToCart = (product: Product) =>
-    setCart((prevCart) => [...prevCart, product]);
-  const cardContextValues = {
-    cart,
-    addToCart,
-  };
   return (
-    <CartContext.Provider value={cardContextValues}>
+    <CartProvider>
       <Routes>
         <Route path="/" element={<Header />}>
           <Route index element={<Catalog />} />
@@ -26,6 +17,6 @@ export function App() {
           <Route path="*" element={<NotFound />} />
         </Route>
       </Routes>
-    </CartContext.Provider>
+    </CartProvider>
   );
 }
