+++ react-context/my-app/src/components/CartContext.tsx
@@ -1,5 +1,6 @@
 import { createContext } from 'react';
 import { type Product } from '../lib';
+import { useState } from 'react';
 
 export type CartValue = {
   cart: Product[];
@@ -12,3 +13,18 @@ const defaultCartValue: CartValue = {
 };
 
 export const CartContext = createContext(defaultCartValue);
+
+export function CartProvider({ children }: { children: React.ReactNode }) {
+  const [cart, setCart] = useState<Product[]>([]);
+  const addToCart = (product: Product) =>
+    setCart((prevCart) => [...prevCart, product]);
+  const cardContextValues = {
+    cart,
+    addToCart,
+  };
+  return (
+    <CartContext.Provider value={cardContextValues}>
+      {children}
+    </CartContext.Provider>
+  );
+}
