+++ react-context/my-app/src/components/useCart.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { CartContext } from './CartContext';
+
+export function useCart() {
+  return useContext(CartContext);
+}
